Add failed environment and setup files
Upload files for attempt to make cadquery work. It doesn't work with this setup.

cadquery supposedly needs Python 3.8 to 3.10 to operate. These files attempt to provide what is needed to install from pip and run cadquery and Jupyterlab in a container that has Python 3.10 installed. The reqs installations go through without error, the container runs with Jupyterlab, but cadquery installation is still throwing vague compatibility errors when I try to call anything from it in the notebook.

The docs suggest using conda to install at this point. Leaving it here for now with the failed pip + docker setup files.